`jake test` does not exit properly.
This causes some major issues, such as `jake deploy` not working. The
tests run fine, but there is a really weird error that happens after
attempting to `rm -rf node_modules/ripple*` (which is a hack). This
started happening randomly, so it seemed easier to remove the need for
such a hack in the first place, vs continuing to figure out the error.

This solution feels better, as well.. and YES, it does use two
instances of node to test now. There is a _slight_ performance hit,
but it is barely noticeable.

GitHub Issue #428