A user should not be able to select UI text.
An example of UI text would be anything in the left and right panels,
and top (if it is being used).

This also gives a better user experience, because you can not
accidentally click and select, which highlights the UI visually.
This can be annoying, since there is no need (ever) to select UI text.