There need only be one jWorkflow file used in the project.
Some time ago, the npm package of jWorkflow was used (for ease) in the
node (unit) test runner, while the one in `thirdparty/jWorkflow` was
being used in Ripple proper, and in the browser runner.

The npm package is now the sole location of jWorkflow in the project,
and the `thirdparty/jWorkflow` (submodule) folder has been removed.