Non WebKit specific vendor prefixes should not be in any CSS files.
This was something that was initially kept, with the goal of someday
running in a number of browsers (and rendering engines).

However, there are a lot of other barriers to running in other browsers,
and when/if the project works on one or more, these prefixes can easily
be added back. In short, YAGNI.