The `dbfs` module belongs in webworks, while `fs` should be initialized.
`rippls/dbfs` was being initialized for every platform, when it is
strictly used for webworks handset/tablet. Additionally, the original
(async) `ripple/fs` module is not even initialized, or used (when it
could of been re-used for cordova's file implementation).

Given this, `dbfs` was moved to `fs` in webworks.core, and it is
initalized on the platform spec initialize method. Additionally, the
fsCache require could be moved to the same place- Yay! `fs` is now
intialized in `lib/ripple`, and can be used (when needed).

On a personal note, `dbfs` has always angered me, being in the top level
of lib- It just did not feel right (and more so like a hack). Doing this
cleans things up a bit, and causes less confusion.