There is no JS test coverage for this project.
Usage: `jake test:cov`

This uses CoverJS: https://github.com/arian/CoverJS

When you run `jake test:cov`, `test` and `lib` code is instrumented and
copied to a `cov` directory in the root of the project, and executed.
When they have completed, a web based report is generated for viewing.

Note: This is a first pass at creating insight into the test coverage
for this project, and could be added onto (or even changed), whichever
makes the most sense. i.e The web based report is hackily changed to
be easier to digest.

Note: This only works for the node based runner (and not `jake btest`).