The first run check UI was not being displayed when warranted.
Due to recent changes in how the platform module is initialized, the
current platform was now being persisted immediately, which is what the
first run check ui plugin uses to determine when to display itself.

This fixes things, but avoids displaying the UI when a platform is being
set via a URL query string param (i.e. ..?enableripple=cordova-2.0.0).

This fixes this GitHub issue:
https://github.com/blackberry/Ripple-UI/issues/465