Cross origin (HTTPS) XHR calls should use SSL between client & server.
Note: This is only when using the cloud based server, and not a local
instance of Ripple's API server.

There was a recent update that got deployed to rippleapi.heroku.com that
adds support for using SSL (in a fashion). Given that, whenever a user
makes a cross origin XHR call, the client (i.e. Ripple's UI) can always
use HTTPS when calling out to Heroku, as you can "piggyback" off their
SSL certs. This also meant the URL needed to be changed to a non mapped
domain name.