There should be some detailed docs on the Chrome extension target.
Previously, there was some basic information on the (deprecated)
`http://rippledocs.tinyhippos.com` site that gave insight into various
ways to install and use the extension. It made sense to create a doc
file directly in the repo, as well as make a reference to it in the main
README.md file, specially around emulating apps hosted from the
`file:///` protocol.

This is also done in the absence of a traditional wiki for the project.