The (hacky) init of the (geo plugin) map needs to be fixed (properly).
With the update to OpenLayers 2.12, it caused a bug that was originally
identified when the feature was fist submitted, but was fixed in a
(rather hacky) way. I'm not sure if it is a fix or some code cleanup in
the new release of OpenLayers, but it is now happening consistently.

The bug is when you enable Ripple with the Geo Location panel closed,
then open it. An exception will be thrown:

    Uncaught TypeError: Cannot read property 'w' of null

However, if you load Ripple and the Geo Location panel is already open,
the map initializes as expected.

This fixes the issue in the way it should of been fixed. As a result,
the visibility check hack can be removed, as well.