Ripple's (compiled) client file should expose select objects on window.
Currently, any objects (or functions) in Ripple's client code, which
includes any thirdparty code, when injected into a web page, would all
be exposed in window, so you would have things like `require` and
`define` in the "global" namespace, which can cause namespace clashes
with an application's code (as well as the other way around).

Now, the _only_ thing that (should) be exposed is the `ripple`
function (which also has a define method on it), which can be used to
require and define modules (ex: via the JS console).

The main reason this was done was to possibly fix this:

https://github.com/blackberry/Ripple-UI/issues/354

However, it is good practise, in general, and should of been done some
time ago, if not, during the compression (build) step.

Note: This required that OpenLayers be updated to v2.12.

Note: A option to not wrap in an auto-invoked function was added so the
browser test runner tests would be able to access thirdparty code (to
mock, etc). This also meant all the manual references could be removed
in favour of the ones in Ripple's client file.