Almond can be replaced with a more lightweight implementation.
Also, it is confusing (and more complex) to support both a `ripple`
and a `require` method (for `lib/client` files), so the more
lightweight implementation (of `require`) was renamed to (only)
expose a method called `ripple`.

i.e. If you are in `lib/client`, you _must_ use `ripple` (and
`ripple.define).

Note: Since we use parts of Cordova (and not just with the addition of
this piece of code), the LICENSE file was updated to mention this.