Emulate a remote (proxied) website using the hosted build target.
You can now (in addition to the `--path` option) enable over a remote
website (app). Ex:

    ripple emulate --remote http://foo.com

Then navigate to (what is logged):

    http://localhost:PORT

Which should load the remote site as it should be.

To enable Ripple's client UI around it, you can add the enableripple
query param:

    http://localhost:PORT?enableripple=true

Note: This is NOT perfect. Some tweaks will be needed.