A mobile website should be sent the proper user agent (server side).
When running a command like so:

    ripple emulate --remote http://foo.com

Then navigate to the logged (local) url:

    http://localhost:PORT?enableripple=true

When enabling for that page, the current userAgent of the current
device (according to client side db) will be sent to the server and
used instead of the browser's.

Note: This does not cover some edge cases, such as having links with
hardcoded host names in them.

Note: Also cleaned up the current navigator.userAgent (client side)
emulation to be more simple (i.e. if userAgent on the current device is
falsy, use original).